Add files via upload
This Python script is aim to get the RMS of the images. The script takes four arguments: receiver, sideband, target, and track, which are used to construct the file name. And open the fits files with astropy.io.fits, assuming that the target image is stored in the primary HDU. For accurate RMS measurement region, a placeholder bounding box is defined. Finally, the script outputs the RMS value and the region box coordinates, formatted for easy parsing in the shell script.