Add files via upload
This shell script automates the imaging process for multiple targets, frequency bands, and sidebands of 200-400GHz Ophiuchus survey using Miriad tasks. Specifically, to achieve the best signal-to-noise ratio (SNR), we used the invert task with natural weighting to perform the inverse Fourier transform for the calibrated visibilities. We then used the clean task to deconvolve the dirty images that were produced from inverse Fourier transform. Finally, we used the restor task to produce the clean images that have Gaussian synthesized beams.